+++ example/android/app/src/main/kotlin/com/qrcodedartscan/qr_code_dart_scan_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.qrcodedartscan.qr_code_dart_scan_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()